+++ dflib/src/main/java/com/nhl/dflib/series/BooleanBaseSeries.java
@@ -2,13 +2,16 @@ package com.nhl.dflib.series;
 
 import com.nhl.dflib.BooleanSeries;
 import com.nhl.dflib.DataFrame;
+import com.nhl.dflib.Index;
 import com.nhl.dflib.IntSeries;
 import com.nhl.dflib.Series;
 import com.nhl.dflib.SeriesGroupBy;
 import com.nhl.dflib.ValueMapper;
 import com.nhl.dflib.ValuePredicate;
+import com.nhl.dflib.ValueToRowMapper;
 import com.nhl.dflib.concat.SeriesConcat;
 import com.nhl.dflib.groupby.SeriesGrouper;
+import com.nhl.dflib.map.Mapper;
 import com.nhl.dflib.sample.Sampler;
 import com.nhl.dflib.series.builder.BooleanAccumulator;
 import com.nhl.dflib.series.builder.IntAccumulator;
@@ -27,6 +30,11 @@ public abstract class BooleanBaseSeries implements BooleanSeries {
     @Override
     public <V> Series<V> map(ValueMapper<Boolean, V> mapper) {
         return new ColumnMappedSeries<>(this, mapper);
+    }
+
+    @Override
+    public DataFrame map(Index resultColumns, ValueToRowMapper<Boolean> mapper) {
+        return Mapper.map(this, resultColumns, mapper);
     }
 
     @Override